requirements.txt
streamlit==1.19.0
SpeechRecognition==3.9.0
google-cloud-language==2.9.0
protobuf==3.20.3


